requirements.txt
Flask
redis
requests
python-dotenv
